Move test utils from elcontracts to general test utils folder
