+++ testing/testing-utils/src/chain_clients.rs
@@ -1,10 +1,12 @@
 use alloy::{
-    primitives::{keccak256, FixedBytes, U256, U8},
+    hex::FromHex,
+    primitives::{address, keccak256, Address, Bytes, FixedBytes, U256, U8},
     sol_types::SolValue,
 };
 use eigen_client_elcontracts::{reader::ELChainReader, writer::ELChainWriter};
 use eigen_common::{get_provider, get_signer};
 use eigen_logging::get_test_logger;
+use std::str::FromStr;
 
 use eigen_client_avsregistry::{reader::AvsRegistryChainReader, writer::AvsRegistryChainWriter};
 use eigen_utils::slashing::{
@@ -26,6 +28,12 @@ use crate::anvil_constants::{
     get_registry_coordinator_address, get_rewards_coordinator_address, get_service_manager_address,
     get_strategy_manager_address, FIRST_ADDRESS, FIRST_PRIVATE_KEY,
 };
+
+/// address for operator used in tests (anvil second address)
+pub const OPERATOR_ADDRESS: Address = address!("70997970C51812dc3A010C7d01b50e0d17dc79C8");
+/// private key for operator used in tests (anvil second private key)
+pub const OPERATOR_PRIVATE_KEY: &str =
+    "0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d";
 
 /// Creates an avs registry chain writer from a given http endpoint and a private key
 pub async fn build_avs_registry_chain_writer(
@@ -106,6 +114,79 @@ pub async fn new_test_writer(http_endpoint: String, private_key: String) -> ELCh
         http_endpoint.clone(),
         private_key,
     )
+}
+
+/// Creates an eigen layer chain writer for M2 quorums (pre-slashing) from a given http endpoint
+pub async fn new_test_writer_preslashing(
+    http_endpoint: String,
+    private_key: String,
+) -> ELChainWriter {
+    let el_chain_reader = build_el_chain_reader(http_endpoint.clone()).await;
+    let strategy_manager = get_strategy_manager_address(http_endpoint.clone()).await;
+    let rewards_coordinator = get_rewards_coordinator_address(http_endpoint.clone()).await;
+    let registry_coordinator = get_registry_coordinator_address(http_endpoint.clone()).await;
+
+    ELChainWriter::new(
+        strategy_manager,
+        rewards_coordinator,
+        None,
+        None,
+        registry_coordinator,
+        el_chain_reader,
+        http_endpoint.clone(),
+        private_key,
+    )
+}
+
+/// Using test data taken from
+/// https://github.com/Layr-Labs/eigenlayer-contracts/blob/a888a1cd1479438dda4b138245a69177b125a973/src/test/test-data/rewardsCoordinator/processClaimProofs_MaxEarnerAndLeafIndices.json
+pub async fn new_test_claim(http_endpoint: &str) -> (FixedBytes<32>, RewardsMerkleClaim) {
+    let el_chain_reader = build_el_chain_reader(http_endpoint.to_string()).await;
+
+    let earner_address = address!("25a1b7322f9796b26a4bec125913b34c292b28d6");
+    let claim = RewardsMerkleClaim {
+            rootIndex: 0,
+            earnerIndex: 7,
+            earnerTreeProof: Bytes::from_hex("4bf5e16eaabbc36964f1e1639808669420f55d60e51adb7e9695b77145c479fd6777be59643947bb24d78e69d6605bf369c515b479f3a8967dd68a97c5bb4a4a262b28002eeb6cbbffb7e79e5741bf2be189a6073440a62fabcd8af4dbda94e3").unwrap(),
+            earnerLeaf: EarnerTreeMerkleLeaf {
+                earner: earner_address,
+                earnerTokenRoot: FixedBytes::from_hex(
+                    "f8e7e20b32aae1d818dcb593b98982841e9a0ed12c161ad603e3ee3948746cba",
+                )
+                .unwrap(),
+            },
+            tokenIndices: vec![7],
+            tokenTreeProofs: vec![
+                Bytes::from_hex("3cd04e8fc6f23812c570fe12292a30bb9e105e00f5913ac4b4938f23e65d8d10e6b1403d58c9d5450952e7d96c81305dad9fb966e8a27d3a42058e3958a0d30033148e91b455542d05deb81b8305b672e742cd3145f7022a0089bad2e6af9173").unwrap(),
+            ],
+            tokenLeaves: vec![TokenTreeMerkleLeaf {
+                token: address!("7fbfdd1dfd80730385aee232cc9f79b8ae12a654"),
+                cumulativeEarnings: U256::from_str("3000000000000000000").unwrap(),
+            }],
+        };
+
+    let root =
+        FixedBytes::from_hex("37550707c80f3d8907c467999730e52127ab89be3f17a5017a3f1ffb73a1445f")
+            .unwrap();
+
+    let key = "ac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80";
+    let rewards_coordinator = IRewardsCoordinator::new(
+        get_rewards_coordinator_address(http_endpoint.to_string()).await,
+        get_signer(key, http_endpoint),
+    );
+    let curr_rewards_calculation_end_timestamp = el_chain_reader
+        .curr_rewards_calculation_end_timestamp()
+        .await
+        .unwrap();
+    let submit_tx = rewards_coordinator
+        .submitRoot(root, curr_rewards_calculation_end_timestamp + 1)
+        .send()
+        .await
+        .unwrap();
+    let submit_status = submit_tx.get_receipt().await.unwrap().status();
+    assert!(submit_status);
+
+    (root, claim)
 }
 
 /// Creates a new claim of amount <cumulative_earnings> in the anvil on the address defined in http_endpoint.
